Blueprint: ssh-pam-tally2
There is no locking policy configured
for SSH failed login attempts in Fuel
for Openstack environments and that is
a major security issue.
Pam_tally2 module should be used in
order to lock user accounts after
certain number of failed ssh login
attempts made to the system,
and unlock it later as well.

Change-Id: I6fd90f582059213e25e555b8468af45d732dc7ca
Signed-off-by: Bogdan Dobrelya <bdobrelia@mirantis.com>