VmWare UI Settings Tab for FuelWeb
VmWare vCenter settings are currently available on Settings Tab
and are only available before the deploy is done. It is very
desirable to have the settings available after deploy.
It is also required to support multiple vCenters for
Compute service. The settings will be always available in
VmWare tab of Cluster page, the functionality will be moved
to a plugin when plugins are available.

Change-Id: Ic8d47d2cc82e5710bd0431b506de93ea8f1884fb
Implements: blueprint vmware-ui-settings