Split fuel-web repository in multiple components
The fuel-web repository currently contains lots of projects which
should be maintained independently: fuel_agent, nailgun,
network_checker, shotgun, tasklib, fuelmenu, and more. As a result,
it's hard to package fuel-web in a proper way, and there's no proper
versionning of each element. Separating them will make it possible
to build each individual component individually. Also, in the
current form, it's very hard to package the fuel-web repository for
distributions.

Change-Id: I2b6151c64b8476781e77e935a3612d8dd5e84918