Introduce component lead for fuel-ui
Now that fuel-ui has been separated into its own repository, we can get
get rid of the special case for it in our team structure, and elect a
component lead for fuel-ui in the Newton cycle. This separation also
allows us to go back to 1:1 matching between fuel-web the repository and
fuel-web the component.

Change-Id: I1346740c2ff9d41eb88af5312b8ddcffedbe5c8b