Deprecate Fuel-wide CL role
As far as we are on our way to make Fuel even
more modular this patch deprecates the role CL
as a Fuel-wide role. Instead Fuel should become
and aggregator of features that provide independent
components. It should be up to a component team
to decide whether they need formal or informal CL.
It also should be up to them how they make architectural
and planning decisions.

Change-Id: Idc35b37711e329a41501246cd082d34eef1e2207